Множественные фиксы всего

+++ code/game/gamemodes/cult/talisman.dm
@@ -1,5 +1,6 @@
 /obj/item/weapon/paper/talisman
 	icon_state = "paper_talisman"
+	var/busy = FALSE
 	var/datum/religion_rites/rite
 	var/datum/religion/religion
 
@@ -7,8 +8,11 @@
 
 /obj/item/weapon/paper/talisman/atom_init(mapload, datum/religion/_religion, datum/religion_rites/_rite)
 	. = ..()
+	RegisterSignal(src, list(COMSIG_OBJ_RESET_RITE), .proc/reset_rite)
+
 	rite = _rite
 	religion = _religion
+	verbs -= /obj/item/weapon/paper/verb/crumple
 
 /obj/item/weapon/paper/talisman/Destroy()
 	rite = null
@@ -18,12 +22,19 @@
 /obj/item/weapon/paper/talisman/update_icon()
 	return
 
+/obj/item/weapon/paper/talisman/proc/reset_rite()
+	busy = FALSE
+
 /obj/item/weapon/paper/talisman/attack_self(mob/living/user)
 	if(!religion?.is_member(user))
 		user.examinate(src)
 		return
+	if(busy)
+		to_chat(user, "<span class='[religion.style_text]'>Талисман уже используется.</span>")
+		return
 
-	user.take_overall_damage(rand(5, 8), rand(1, 5))
+	busy = TRUE
+	user.take_overall_damage(rand(1, 5), rand(1, 5))
 	if(rite?.perform_rite(user, src))
 		if(disposable)
 			qdel(src)
